initial support of 1.8.21

+++ src/types/get_file_downloaded_prefix_size.rs
@@ -17,7 +17,7 @@ pub struct GetFileDownloadedPrefixSize {
     /// Offset from which downloaded prefix size needs to be calculated
 
     #[serde(default)]
-    offset: i32,
+    offset: i64,
 
     #[serde(rename(serialize = "@type"))]
     td_type: String,
@@ -53,7 +53,7 @@ impl GetFileDownloadedPrefixSize {
         self.file_id
     }
 
-    pub fn offset(&self) -> i32 {
+    pub fn offset(&self) -> i64 {
         self.offset
     }
 }
@@ -76,7 +76,7 @@ impl GetFileDownloadedPrefixSizeBuilder {
         self
     }
 
-    pub fn offset(&mut self, offset: i32) -> &mut Self {
+    pub fn offset(&mut self, offset: i64) -> &mut Self {
         self.inner.offset = offset;
         self
     }
